--- front/app/components/admin/ExportMenu/messages.ts
@@ -1,32 +1,0 @@
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  fromFilter: {
-    id: 'app.components.admin.ExportMenu.FileName.fromFilter',
-    defaultMessage: 'from',
-  },
-  untilFilter: {
-    id: 'app.components.admin.ExportMenu.FileName.untilFilter',
-    defaultMessage: 'until',
-  },
-  projectFilter: {
-    id: 'app.components.admin.ExportMenu.FileName.projectFilter',
-    defaultMessage: 'project',
-  },
-  groupFilter: {
-    id: 'app.components.admin.ExportMenu.FileName.groupFilter',
-    defaultMessage: 'group',
-  },
-  topicFilter: {
-    id: 'app.components.admin.ExportMenu.FileName.topicFilter',
-    defaultMessage: 'topic',
-  },
-  downloadAsImage: {
-    id: 'app.components.admin.ExportMenu.downloadAsImage',
-    defaultMessage: 'Download as image',
-  },
-  downloadXlsx: {
-    id: 'app.components.admin.ExportMenu.downloadXlsx',
-    defaultMessage: 'Download Excel',
-  },
-});
