+++ front/app/components/admin/PostManager/PostManager.test.tsx
@@ -14,7 +14,7 @@ jest.mock('resources/GetInitiativeStatuses', () => 'GetInitiativeStatuses');
 jest.mock('resources/GetIdeas', () => 'GetIdeas');
 jest.mock('resources/GetInitiatives', () => 'GetInitiatives');
 jest.mock('./components/ActionBar', () => 'ActionBar');
-jest.mock('./components/ExportMenu', () => 'ExportMenu');
+jest.mock('./components/ReportExportMenu', () => 'ReportExportMenu');
 jest.mock('./components/FilterSidebar', () => 'FilterSidebar');
 jest.mock('./components/PostTable', () => 'PostTable');
 jest.mock('./components/InfoSidebar', () => 'InfoSidebar');
