+++ front/app/components/admin/ReportExportMenu/messages.ts
@@ -1,0 +1,32 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  fromFilter: {
+    id: 'app.components.admin.ReportExportMenu.FileName.fromFilter',
+    defaultMessage: 'from',
+  },
+  untilFilter: {
+    id: 'app.components.admin.ReportExportMenu.FileName.untilFilter',
+    defaultMessage: 'until',
+  },
+  projectFilter: {
+    id: 'app.components.admin.ReportExportMenu.FileName.projectFilter',
+    defaultMessage: 'project',
+  },
+  groupFilter: {
+    id: 'app.components.admin.ReportExportMenu.FileName.groupFilter',
+    defaultMessage: 'group',
+  },
+  topicFilter: {
+    id: 'app.components.admin.ReportExportMenu.FileName.topicFilter',
+    defaultMessage: 'topic',
+  },
+  downloadAsImage: {
+    id: 'app.components.admin.ReportExportMenu.downloadAsImage',
+    defaultMessage: 'Download as image',
+  },
+  downloadXlsx: {
+    id: 'app.components.admin.ReportExportMenu.downloadXlsx',
+    defaultMessage: 'Download Excel',
+  },
+});
