+++ front/app/translations/de-DE.json
@@ -470,13 +470,13 @@
   "app.components.VoteControl.votingPossibleLater": "Die Stimmabgabe beginnt am {enabledFromDate}\n",
   "app.components.admin.DateRangePicker.endDatePlaceholder": "Enddatum",
   "app.components.admin.DateRangePicker.startDatePlaceholder": "Start date",
-  "app.components.admin.ExportMenu.FileName.fromFilter": "",
-  "app.components.admin.ExportMenu.FileName.groupFilter": "",
-  "app.components.admin.ExportMenu.FileName.projectFilter": "",
-  "app.components.admin.ExportMenu.FileName.topicFilter": "",
-  "app.components.admin.ExportMenu.FileName.untilFilter": "",
-  "app.components.admin.ExportMenu.downloadAsImage": "",
-  "app.components.admin.ExportMenu.downloadXlsx": "",
+  "app.components.admin.ReportExportMenu.FileName.fromFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.groupFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.projectFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.topicFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.untilFilter": "",
+  "app.components.admin.ReportExportMenu.downloadAsImage": "",
+  "app.components.admin.ReportExportMenu.downloadXlsx": "",
   "app.components.admin.FormikSubmitWrapper.buttonSave": "Speichern",
   "app.components.admin.FormikSubmitWrapper.buttonSuccess": "Gespeichert!",
   "app.components.admin.FormikSubmitWrapper.messageError": "Es ist ein Fehler beim Speichern aufgetreten",
