+++ front/app/translations/en.json
@@ -470,13 +470,13 @@
   "app.components.VoteControl.votingPossibleLater": "Voting will start on {enabledFromDate}",
   "app.components.admin.DateRangePicker.endDatePlaceholder": "End Date",
   "app.components.admin.DateRangePicker.startDatePlaceholder": "Start date",
-  "app.components.admin.ExportMenu.FileName.fromFilter": "from",
-  "app.components.admin.ExportMenu.FileName.groupFilter": "group",
-  "app.components.admin.ExportMenu.FileName.projectFilter": "project",
-  "app.components.admin.ExportMenu.FileName.topicFilter": "topic",
-  "app.components.admin.ExportMenu.FileName.untilFilter": "until",
-  "app.components.admin.ExportMenu.downloadAsImage": "Download as image",
-  "app.components.admin.ExportMenu.downloadXlsx": "Download Excel",
+  "app.components.admin.ReportExportMenu.FileName.fromFilter": "from",
+  "app.components.admin.ReportExportMenu.FileName.groupFilter": "group",
+  "app.components.admin.ReportExportMenu.FileName.projectFilter": "project",
+  "app.components.admin.ReportExportMenu.FileName.topicFilter": "topic",
+  "app.components.admin.ReportExportMenu.FileName.untilFilter": "until",
+  "app.components.admin.ReportExportMenu.downloadAsImage": "Download as image",
+  "app.components.admin.ReportExportMenu.downloadXlsx": "Download Excel",
   "app.components.admin.FormikSubmitWrapper.buttonSave": "Save",
   "app.components.admin.FormikSubmitWrapper.buttonSuccess": "Success",
   "app.components.admin.FormikSubmitWrapper.messageError": "There was an error during saving",
