+++ front/app/translations/es-ES.json
@@ -470,13 +470,13 @@
   "app.components.VoteControl.votingPossibleLater": "La votación comenzará el {enabledFromDate}",
   "app.components.admin.DateRangePicker.endDatePlaceholder": "Fecha de finalización",
   "app.components.admin.DateRangePicker.startDatePlaceholder": "Fecha de inicio",
-  "app.components.admin.ExportMenu.FileName.fromFilter": "",
-  "app.components.admin.ExportMenu.FileName.groupFilter": "",
-  "app.components.admin.ExportMenu.FileName.projectFilter": "",
-  "app.components.admin.ExportMenu.FileName.topicFilter": "",
-  "app.components.admin.ExportMenu.FileName.untilFilter": "",
-  "app.components.admin.ExportMenu.downloadAsImage": "",
-  "app.components.admin.ExportMenu.downloadXlsx": "",
+  "app.components.admin.ReportExportMenu.FileName.fromFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.groupFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.projectFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.topicFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.untilFilter": "",
+  "app.components.admin.ReportExportMenu.downloadAsImage": "",
+  "app.components.admin.ReportExportMenu.downloadXlsx": "",
   "app.components.admin.FormikSubmitWrapper.buttonSave": "Guardar",
   "app.components.admin.FormikSubmitWrapper.buttonSuccess": "Éxito",
   "app.components.admin.FormikSubmitWrapper.messageError": "Se produjo un error al guardar",
