+++ front/app/translations/fr-BE.json
@@ -470,13 +470,13 @@
   "app.components.VoteControl.votingPossibleLater": "Le vote commencera le {enabledFromDate}",
   "app.components.admin.DateRangePicker.endDatePlaceholder": "Date de fin",
   "app.components.admin.DateRangePicker.startDatePlaceholder": "Date de début",
-  "app.components.admin.ExportMenu.FileName.fromFilter": "",
-  "app.components.admin.ExportMenu.FileName.groupFilter": "",
-  "app.components.admin.ExportMenu.FileName.projectFilter": "",
-  "app.components.admin.ExportMenu.FileName.topicFilter": "",
-  "app.components.admin.ExportMenu.FileName.untilFilter": "",
-  "app.components.admin.ExportMenu.downloadAsImage": "",
-  "app.components.admin.ExportMenu.downloadXlsx": "",
+  "app.components.admin.ReportExportMenu.FileName.fromFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.groupFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.projectFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.topicFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.untilFilter": "",
+  "app.components.admin.ReportExportMenu.downloadAsImage": "",
+  "app.components.admin.ReportExportMenu.downloadXlsx": "",
   "app.components.admin.FormikSubmitWrapper.buttonSave": "Enregistrer",
   "app.components.admin.FormikSubmitWrapper.buttonSuccess": "Enregistré !",
   "app.components.admin.FormikSubmitWrapper.messageError": "Une erreur est survenue lors de la sauvegarde",
