+++ front/app/translations/mi-NZ.json
@@ -937,13 +937,13 @@
   "app.containers.Admin.projects.all.deleteProjectButton": "Delete",
   "app.containers.Admin.projects.all.deleteProjectConfirmation": "Are you sure you want to delete this project? This cannot be undone.",
   "app.containers.Admin.projects.all.deleteProjectError": "There was an error deleting this project, please try again later.",
-  "app.containers.AdminPage.DashboardPage.ExportMenu.FileName.fromFilter": "from",
-  "app.containers.AdminPage.DashboardPage.ExportMenu.FileName.groupFilter": "group",
-  "app.containers.AdminPage.DashboardPage.ExportMenu.FileName.projectFilter": "project",
-  "app.containers.AdminPage.DashboardPage.ExportMenu.FileName.topicFilter": "topic",
-  "app.containers.AdminPage.DashboardPage.ExportMenu.FileName.untilFilter": "until",
-  "app.containers.AdminPage.DashboardPage.ExportMenu.downloadAsImage": "Download as image",
-  "app.containers.AdminPage.DashboardPage.ExportMenu.downloadXlsx": "Download Excel",
+  "app.containers.AdminPage.DashboardPage.ReportExportMenu.FileName.fromFilter": "from",
+  "app.containers.AdminPage.DashboardPage.ReportExportMenu.FileName.groupFilter": "group",
+  "app.containers.AdminPage.DashboardPage.ReportExportMenu.FileName.projectFilter": "project",
+  "app.containers.AdminPage.DashboardPage.ReportExportMenu.FileName.topicFilter": "topic",
+  "app.containers.AdminPage.DashboardPage.ReportExportMenu.FileName.untilFilter": "until",
+  "app.containers.AdminPage.DashboardPage.ReportExportMenu.downloadAsImage": "Download as image",
+  "app.containers.AdminPage.DashboardPage.ReportExportMenu.downloadXlsx": "Download Excel",
   "app.containers.AdminPage.DashboardPage.Report.participants": "participants",
   "app.containers.AdminPage.DashboardPage.Report.totalUsers": "total on the platform",
   "app.containers.AdminPage.DashboardPage._blank": "unknown",
