+++ front/app/translations/mi.json
@@ -470,13 +470,13 @@
   "app.components.VoteControl.votingPossibleLater": "Voting will start on {enabledFromDate}",
   "app.components.admin.DateRangePicker.endDatePlaceholder": "End Date",
   "app.components.admin.DateRangePicker.startDatePlaceholder": "Start date",
-  "app.components.admin.ExportMenu.FileName.fromFilter": "",
-  "app.components.admin.ExportMenu.FileName.groupFilter": "",
-  "app.components.admin.ExportMenu.FileName.projectFilter": "",
-  "app.components.admin.ExportMenu.FileName.topicFilter": "",
-  "app.components.admin.ExportMenu.FileName.untilFilter": "",
-  "app.components.admin.ExportMenu.downloadAsImage": "",
-  "app.components.admin.ExportMenu.downloadXlsx": "",
+  "app.components.admin.ReportExportMenu.FileName.fromFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.groupFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.projectFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.topicFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.untilFilter": "",
+  "app.components.admin.ReportExportMenu.downloadAsImage": "",
+  "app.components.admin.ReportExportMenu.downloadXlsx": "",
   "app.components.admin.FormikSubmitWrapper.buttonSave": "Save",
   "app.components.admin.FormikSubmitWrapper.buttonSuccess": "Success",
   "app.components.admin.FormikSubmitWrapper.messageError": "There was an error during saving",
