+++ front/app/translations/nl-BE.json
@@ -470,13 +470,13 @@
   "app.components.VoteControl.votingPossibleLater": "Stemmen begint op {enabledFromDate}",
   "app.components.admin.DateRangePicker.endDatePlaceholder": "Einddatum",
   "app.components.admin.DateRangePicker.startDatePlaceholder": "Begint op",
-  "app.components.admin.ExportMenu.FileName.fromFilter": "",
-  "app.components.admin.ExportMenu.FileName.groupFilter": "",
-  "app.components.admin.ExportMenu.FileName.projectFilter": "",
-  "app.components.admin.ExportMenu.FileName.topicFilter": "",
-  "app.components.admin.ExportMenu.FileName.untilFilter": "",
-  "app.components.admin.ExportMenu.downloadAsImage": "",
-  "app.components.admin.ExportMenu.downloadXlsx": "",
+  "app.components.admin.ReportExportMenu.FileName.fromFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.groupFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.projectFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.topicFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.untilFilter": "",
+  "app.components.admin.ReportExportMenu.downloadAsImage": "",
+  "app.components.admin.ReportExportMenu.downloadXlsx": "",
   "app.components.admin.FormikSubmitWrapper.buttonSave": "Opslaan",
   "app.components.admin.FormikSubmitWrapper.buttonSuccess": "Opgeslagen!",
   "app.components.admin.FormikSubmitWrapper.messageError": "Er is een fout opgetreden tijdens het opslaan",
