+++ front/app/translations/ro-RO.json
@@ -470,13 +470,13 @@
   "app.components.VoteControl.votingPossibleLater": "Votarea va începe pe {enabledFromDate}",
   "app.components.admin.DateRangePicker.endDatePlaceholder": "Data de încheiere",
   "app.components.admin.DateRangePicker.startDatePlaceholder": "Data de începere",
-  "app.components.admin.ExportMenu.FileName.fromFilter": "",
-  "app.components.admin.ExportMenu.FileName.groupFilter": "",
-  "app.components.admin.ExportMenu.FileName.projectFilter": "",
-  "app.components.admin.ExportMenu.FileName.topicFilter": "",
-  "app.components.admin.ExportMenu.FileName.untilFilter": "",
-  "app.components.admin.ExportMenu.downloadAsImage": "",
-  "app.components.admin.ExportMenu.downloadXlsx": "",
+  "app.components.admin.ReportExportMenu.FileName.fromFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.groupFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.projectFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.topicFilter": "",
+  "app.components.admin.ReportExportMenu.FileName.untilFilter": "",
+  "app.components.admin.ReportExportMenu.downloadAsImage": "",
+  "app.components.admin.ReportExportMenu.downloadXlsx": "",
   "app.components.admin.FormikSubmitWrapper.buttonSave": "Salvați",
   "app.components.admin.FormikSubmitWrapper.buttonSuccess": "Succes",
   "app.components.admin.FormikSubmitWrapper.messageError": "A apărut o eroare în timpul salvării",
