+++ front/app/translations/sr-CS.json
@@ -957,13 +957,13 @@
   "app.containers.Admin.projects.all.deleteProjectButton": "Izbriši",
   "app.containers.Admin.projects.all.deleteProjectConfirmation": "Da li ste sigurni da želite da obrišete ovaj projekat? Ova aktivnost se ne može poništiti.",
   "app.containers.Admin.projects.all.deleteProjectError": "Došlo je do greške prilikom brisanja ovog projekta, molimo vas pokušajte kasnije.",
-  "app.containers.AdminPage.DashboardPage.ExportMenu.FileName.fromFilter": "od",
-  "app.containers.AdminPage.DashboardPage.ExportMenu.FileName.groupFilter": "grupa",
-  "app.containers.AdminPage.DashboardPage.ExportMenu.FileName.projectFilter": "projekat",
-  "app.containers.AdminPage.DashboardPage.ExportMenu.FileName.topicFilter": "tema",
-  "app.containers.AdminPage.DashboardPage.ExportMenu.FileName.untilFilter": "dok",
-  "app.containers.AdminPage.DashboardPage.ExportMenu.downloadAsImage": "Preuzmi kao sliku",
-  "app.containers.AdminPage.DashboardPage.ExportMenu.downloadXlsx": "Preuzmi Excel",
+  "app.containers.AdminPage.DashboardPage.ReportExportMenu.FileName.fromFilter": "od",
+  "app.containers.AdminPage.DashboardPage.ReportExportMenu.FileName.groupFilter": "grupa",
+  "app.containers.AdminPage.DashboardPage.ReportExportMenu.FileName.projectFilter": "projekat",
+  "app.containers.AdminPage.DashboardPage.ReportExportMenu.FileName.topicFilter": "tema",
+  "app.containers.AdminPage.DashboardPage.ReportExportMenu.FileName.untilFilter": "dok",
+  "app.containers.AdminPage.DashboardPage.ReportExportMenu.downloadAsImage": "Preuzmi kao sliku",
+  "app.containers.AdminPage.DashboardPage.ReportExportMenu.downloadXlsx": "Preuzmi Excel",
   "app.containers.AdminPage.DashboardPage.Report.participants": "učesnici",
   "app.containers.AdminPage.DashboardPage.Report.totalUsers": "ukupno na platformi",
   "app.containers.AdminPage.DashboardPage._blank": "nepoznato",
